Force into to shape of x0 Force shape[1] of LinearOperator to be int when using it to create used x0 in __truediv__ method. This is required because cupy does not accept np.int type to the passed shape of an array creation.

+++ pylops/LinearOperator.py
@@ -233,7 +233,8 @@ class LinearOperator(spLinearOperator):
             else:
                 # cupy backend
                 ncp = get_array_module(y)
-                xest = cgls(self, y, x0=ncp.zeros(self.shape[1]),
+                xest = cgls(self, y,
+                            x0=ncp.zeros(int(self.shape[1]), dtype=self.dtype),
                             niter=niter)[0]
         return xest
